pyproject.toml
[tool.poetry]
name = "anything"
version = "0.1.0"
description = "try grpc"
authors = ["Poe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.65.1"
grpcio-tools = "^1.65.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
